Repairing an error in the 5th <hr> element
For some reason, the <hr> element, in the 5th flex children was not as thick, as the other flex children's <hr> element. To battle that I gave an id of "dk-wassup-with-this", targeted it in the styles.css file, and have succesfully achieved, so that the <hr> in the 5th flex children element is as thick as the other flex childrens' <hr> element.